% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_credit.R
\name{get_credit}
\alias{get_credit}
\title{Get basemap tiles attribution}
\usage{
get_credit(provider)
}
\arguments{
\item{provider}{provider name}
}
\description{
Get the attribution of map tiles.
}
\examples{
get_credit("OpenStreetMap")
}
