% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_providers.R
\name{get_providers}
\alias{get_providers}
\title{Providers}
\usage{
get_providers()
}
\value{
A list of is returned.
}
\description{
List of builtin providers with their name, URL, subdomains and
attribution text.
}
\examples{
get_providers()
}
