% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tiles.R
\name{mp_get_tiles}
\alias{mp_get_tiles}
\title{Get basemap tiles from map servers}
\usage{
mp_get_tiles(
  x,
  type = "OpenStreetMap",
  zoom = NULL,
  crop = FALSE,
  verbose = FALSE,
  apikey = NA,
  cachedir = FALSE,
  forceDownload = FALSE
)
}
\arguments{
\item{x}{an sf or sfc object.}

\item{type}{the tile server from which to get the map. See Details for providers.
For other sources use a list: type = list(src = "name of the source" ,
q = "tiles address", sub = "subdomains", cit = "how to cite the tiles"). See Examples.}

\item{zoom}{the zoom level.}

\item{crop}{TRUE if results should be cropped to the specified x extent,
FALSE otherwise. If x is an sf object with one POINT, crop is set to FALSE.}

\item{verbose}{if TRUE, tiles filepaths, zoom level and citation are displayed.}

\item{apikey}{Needed for Thunderforest maps.}

\item{cachedir}{name of a directory used to cache tiles. If TRUE, places a
'tile.cache' folder in the working directory. If FALSE, tiles are only
cached in \link[base:tempdir]{tempdir}.}

\item{forceDownload}{if TRUE, cached tiles are downloaded again.}
}
\value{
A SpatRaster is returned.
}
\description{
Get map tiles based on a spatial object extent. Maps can be
fetched from various open map servers.
}
\details{
Zoom levels are described on the OpenStreetMap wiki:
\url{http://wiki.openstreetmap.org/wiki/Zoom_levels}. \cr\cr
}
\examples{
\dontrun{
library(sf)
}
}
