% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tiles_attribution.R
\name{mp_get_tiles_attribution}
\alias{mp_get_tiles_attribution}
\title{Get basemap tiles attribution}
\usage{
mp_get_tiles_attribution(provider)
}
\arguments{
\item{provider}{provider name}
}
\description{
Get the attribution of map tiles.
}
\examples{
mp_get_tiles_attribution("OpenStreetMap")
}
