% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_provider.R
\name{create_provider}
\alias{create_provider}
\title{Create a new tile provider}
\usage{
create_provider(name, url, sub = NA, citation)
}
\arguments{
\item{name}{name of the provider}

\item{url}{url of the provider. The url must contain {x}, {y} and {z} placeholder}

\item{sub}{sub-domains}

\item{citation}{attribution text of the provider}
}
\value{
a list is returned. This list can be used by get_tiles()
}
\description{
Create a new tile provider
}
