% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{maptiles}
\alias{maptiles}
\title{Download and Display Map Tiles}
\description{
To create maps from tiles, maptiles downloads, composes and displays tiles
from a large number of providers (e.g. OpenStreetMap, Stamen, Esri, CARTO,
or Thunderforest).
}
